# colorcontrast: optimize code and remove SSE