Eliminating the cacheline fills for the output buffer cuts the runtime
enough to make the plain code slightly faster than the SSE codepath.
```
Thr     SSE     new-plain
1     10.57ms   9.95ms
2      6.19ms   5.99ms
4      4.82ms   4.72ms
6      4.30ms   4.29ms
8      3.92ms   3.88ms
16     4.30ms   4.17ms
32     4.53ms   4.47ms
```
Verified pixel-identical output for integration test 0038.