@jenshannoschwalm 2023-02-19
Confirmed results on 8core intel ...